Add StatDefault to configuration.
StatDefault may be set to one of the following values:

Prior - The stat starts at 0 and is never reset. If the stat is has no value in CloudWatch, the prior value is reported. This was the existing behavior before this was introduced, and is now the default behavior unless otherwise configured.

Zero - The stat starts at 0 and resets back to zero if the stat has no value in CloudWatch.

NaN - The stat starts as not-a-number (NaN) and resets back to NaN if the stat has no value in CloudWatch.